Update for 5.10-android12
Completely removed configs are not available on 5.10 ACK

Applying some or maybe any of the commented configs, builds but does not boot. I only tested CONFIG_CGROUP_DEVICE. Most probably breaking the ABI and not loading the modules as a result. The device is stuck in ramdisk and reboots after 60sec timeout.

Waydroid is not tested, so we might still miss some configs for that.